AP_Motors:Heli - change param names for governor to be compatible with next-gen GCS

+++ libraries/AP_Motors/AP_MotorsHeli_Quad.h
@@ -37,7 +37,7 @@ public:
     void output_test(uint8_t motor_seq, int16_t pwm) override;
 
     // set_rpm - for rotor speed governor
-    void set_rpm(int16_t rotor_rpm) override;
+    void set_rpm(float rotor_rpm) override;
 
     // set_desired_rotor_speed - sets target rotor speed as a number from 0 ~ 1000
     void set_desired_rotor_speed(float desired_speed) override;
